Testing sin impuesto

+++ src/main/java/com/TP/TallerMecanico/entidad/Marca.java
@@ -3,7 +3,6 @@ package com.TP.TallerMecanico.entidad;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.*;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -28,7 +27,7 @@ public class Marca implements Serializable{
     @NotNull(message = "El campo no debe estar vacio")
     @DecimalMin(value = "0.0", inclusive = false, message = "El impuesto debe ser mayor que cero")
     @DecimalMax(value = "100.0", inclusive = true, message = "El impuesto no puede ser mayor que 100%")
-    private BigDecimal impuesto;
+    private Double impuesto;
     
     // @NotNull(message = "El campo no debe estar vacio")
     // @DecimalMin(value = "0.0", inclusive = false, message = "El impuesto debe ser mayor que cero")
